Make zap reports work in cloud env
Introduce a transitional directory to copy zap reports thus changing
user:group ownership of report files, directories and support files.
We then recursivly update permissions in order to be able to traverse
the nested directories and modify markup and styling.

The orchestrator is thus able to traverse the report files and
directories and also able to delete.

The cloud instances periodically run a cron job to modify the
user:group of the report files, directories and support files in
order for us to attempt to delete immediately before we request zap
generate reports.